Create VIDEO_LIST_FILE.txt
Place this file in the same directory containing the CONFIG.py and FYTDL.py files. Each YouTube video link or other video link must be placed on its own line. This repository is not affiliated with the youtube-dl test video that is linked in this file; the link is only listed as example.