Update FYTDL.py
This script leverages the power of youtube-dl and automates the hashing, logging, and results reporting processes. This repository is not affiliated with the contributors of or the repository named youtube-dl. This repository is dependent on the existence of the youtube-dl program which is required in for the script to perform properly; youtube-dl MUST exist on the host machine in order for this script to work.